@@ -1,19 +1,15 @@
 name: list_bloc
 description: A BLoC library for loading data from an api end-point with filtering and pagination.
-version: 1.0.2
+version: 2.0.1
 homepage: https://github.com/apexlabs-ai/list_bloc/
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  bloc: ^6.0.1
-  equatable: ^1.2.2
-#  built_value: ^7.1.0
-
+  bloc: ^7.0.0
+  equatable: ^2.0.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
-#  build_runner: ^1.10.0
-#  built_value_generator: ^7.1.0
+  pedantic: ^1.11.0
+  test: ^1.16.8
